+++ src/Trixi.jl
@@ -72,7 +72,8 @@ using SummationByPartsOperators: AbstractDerivativeOperator,
 import SummationByPartsOperators: integrate, semidiscretize,
                                   left_boundary_weight, right_boundary_weight
 @reexport using SummationByPartsOperators:
-  SummationByPartsOperators, derivative_operator, periodic_derivative_operator
+  SummationByPartsOperators, derivative_operator, periodic_derivative_operator,
+  upwind_operators
 
 # DGMulti solvers
 @reexport using StartUpDG: StartUpDG, Polynomial, SBP, Line, Tri, Quad, Hex, Tet
@@ -132,7 +133,8 @@ export AcousticPerturbationEquations2D,
        LinearScalarAdvectionEquation1D, LinearScalarAdvectionEquation2D, LinearScalarAdvectionEquation3D,
        InviscidBurgersEquation1D,
        LatticeBoltzmannEquations2D, LatticeBoltzmannEquations3D,
-       ShallowWaterEquations1D, ShallowWaterEquations2D, LinearizedEulerEquations2D, MaxwellEquations2D
+       ShallowWaterEquations1D, ShallowWaterEquations2D,
+       LinearizedEulerEquations2D, MaxwellEquations2D
 
 export LaplaceDiffusion2D,
        CompressibleNavierStokesDiffusion2D
@@ -152,7 +154,11 @@ export flux, flux_central, flux_lax_friedrichs, flux_hll, flux_hllc, flux_hlle, 
        FluxLMARS,
        FluxRotated,
        flux_shima_etal_turbo, flux_ranocha_turbo,
-       FluxHydrostaticReconstruction
+       FluxHydrostaticReconstruction,
+       FluxUpwind
+
+export splitting_steger_warming, splitting_vanleer_haenel,
+       splitting_coirier_vanleer, splitting_lax_friedrichs
 
 export initial_condition_constant,
        initial_condition_gauss,
@@ -184,11 +190,14 @@ export TreeMesh, StructuredMesh, UnstructuredMesh2D, P4estMesh
 
 export DG,
        DGSEM, LobattoLegendreBasis,
+       FDSBP,
        VolumeIntegralWeakForm, VolumeIntegralStrongForm,
        VolumeIntegralFluxDifferencing,
        VolumeIntegralPureLGLFiniteVolume,
        VolumeIntegralShockCapturingHG, IndicatorHennemannGassner,
+       VolumeIntegralUpwind,
        SurfaceIntegralWeakForm, SurfaceIntegralStrongForm,
+       SurfaceIntegralUpwind,
        MortarL2
 
 export nelements, nnodes, nvariables,